use double instead of int to compute the x, y, width and height to pass to
2007-06-18  Vincent Untz  <vuntz@gnome.org>

	* libwnck/pager.c: (wnck_pager_draw_workspace): use double instead of
	int to compute the x, y, width and height to pass to cairo. Improves
	the look of the pager a bit.

svn path=/trunk/; revision=1332

+++ libwnck/pager.c
@@ -974,7 +974,7 @@ wnck_pager_draw_workspace (WnckPager    *pager,
     }
   else if (active_space)
     {
-      int vx, vy, vw, vh; /* viewport */
+      double vx, vy, vw, vh; /* viewport */
       double width_ratio, height_ratio;
       cairo_t *cr;
